moved all the weights calc into a different utils file
